[refactor] UserInformation 위치 -> auth
